rename to command-security
