+++ apps/server/package.json
@@ -23,7 +23,7 @@
     "@ai-sdk/anthropic": "^1.2.12",
     "@ai-sdk/openai": "^1.3.23",
     "@repo/db": "*",
-    "@repo/security": "*",
+    "@repo/command-security": "*",
     "@repo/types": "*",
     "@types/dotenv": "^8.2.3",
     "ai": "^4.3.19",
