+++ packages/command-security/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@repo/security",
+  "name": "@repo/command-security",
   "version": "0.0.0",
   "private": true,
   "main": "./dist/index.js",
