+++ packages/command-security/src/command-security.ts
