+++ packages/command-security/src/index.ts
