+++ packages/command-security/tsconfig.json
